requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
xgboost>=1.7.0
opencv-python>=4.7.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
scipy>=1.10.0
scikit-image>=0.20.0